refactor : 예외 처리 강화 및 클래스 분리

+++ src/main/java/org/ezcode/codetest/domain/problem/model/ProblemInfo.java
@@ -12,7 +12,7 @@ public record ProblemInfo(
 	List<Testcase> testcaseList
 
 ) {
-	public double getTimeLimit() {
+	public long getTimeLimit() {
 		return this.problem.getTimeLimit();
 	}
